+++ modules/async_utils/async_main/src/lib.rs
@@ -13,7 +13,7 @@ pub fn async_main(_args: TokenStream, item: TokenStream) -> TokenStream {
         #[used]
         #[no_mangle]
         static ASYNC_MAIN: fn() -> BoxFut = keep_name;
-        type BoxFut = Pin<Box<dyn Future<Output = i32> + 'static>>;
+        type BoxFut = Pin<Box<dyn Future<Output = isize> + 'static>>;
 
         fn keep_name() -> BoxFut {
             Box::pin(async { #block })
